+++ src/commands/next.ts
@@ -2,6 +2,7 @@ import { Command, flags } from '@oclif/command'
 
 import consts from '../common/consts'
 import * as extract from '../common/extract'
+import * as format from '../common/format'
 import glob from '../common/glob'
 
 export default class Next extends Command {
@@ -19,11 +20,31 @@ export default class Next extends Command {
 
   static flags = {
     help: flags.help({ char: 'h' }),
+    format: flags.string({
+      char: 'f',
+      description: 'Specify output format',
+      default: format.Output.markdown,
+      multiple: false,
+      options: [
+        // List output options.
+        format.Output.markdown,
+        format.Output.slack,
+      ],
+      required: false,
+    }),
+    only: flags.boolean({
+      char: 'o',
+      description: 'Only show the changes that are included in given version',
+      default: false,
+      required: false,
+    }),
   }
 
   static args = []
 
   async run() {
+    const { flags } = this.parse(Next)
+
     const targets = glob(consts.defaultMatcher)
 
     if (targets.length === 0) {
@@ -47,13 +68,22 @@ export default class Next extends Command {
     this.log('')
 
     outs.forEach((o) => {
-      this.log('##', o.target, '\n')
+      const header = format.header({
+        format: flags.format as any,
+        value: o.target,
+      })
 
       if (o.messages.length === 0) {
+        if (flags.only) {
+          return
+        }
+
+        this.log(header, '\n')
         this.log(consts.noChanges, consts.postLines)
         return
       }
 
+      this.log(header, '\n')
       this.log(o.messages.join('\n'), consts.postLines)
     })
   }